+++ src/main/java/com/example/digitalimage/model/dao/ArticleContentMapper.java
@@ -3,13 +3,13 @@ package com.example.digitalimage.model.dao;
 import com.example.digitalimage.model.entity.ArticleContent;
 
 public interface ArticleContentMapper {
-    int deleteByPrimaryKey(Integer artId);
+    int deleteByPrimaryKey(Long artId);
 
     int insert(ArticleContent record);
 
     int insertSelective(ArticleContent record);
 
-    ArticleContent selectByPrimaryKey(Integer artId);
+    ArticleContent selectByPrimaryKey(Long artId);
 
     int updateByPrimaryKeySelective(ArticleContent record);
